[refactor] 리팩토링 마무리
