+++ src/main/java/dev/woori/wooriLearn/domain/auth/jwt/JwtIssuer.java
@@ -1,7 +1,7 @@
 package dev.woori.wooriLearn.domain.auth.jwt;
 
 import dev.woori.wooriLearn.config.jwt.TokenInfo;
-import dev.woori.wooriLearn.domain.user.entity.Role;
+import dev.woori.wooriLearn.domain.auth.entity.Role;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
